requirements.txt
# PydanticAI Integration Requirements
pydantic-ai[openai,anthropic,gemini]==0.0.14
pydantic==2.10.4
fastapi==0.115.6
uvicorn[standard]==0.32.1
google-cloud-aiplatform==1.78.0
google-cloud-pubsub==2.30.1
google-cloud-functions==1.19.2
google-cloud-firestore==2.20.0
redis==5.2.1
asyncio-mqtt==0.16.2
websockets==14.1
loguru==0.7.3
python-multipart==0.0.20
httpx==0.28.1
orjson==3.10.12
structlog==24.4.0
psutil==6.1.0

# Trading specific
numpy==2.2.1
pandas==2.2.3
scipy==1.14.1
scikit-learn==1.6.0
ta-lib==0.4.32
yfinance==0.2.50
alpha-vantage==2.3.1
hyperliquid>=0.1.9
hyperliquid-python-sdk>=0.2.0

# Development
pytest==8.3.4
pytest-asyncio==0.25.0
black==24.10.0
isort==5.13.2
mypy==1.13.0

# Core Service Dependencies (Added/Ensured)
python-dotenv>=1.0.0
pyautogen>=0.2.0
autogen-agentchat>=0.2.0
crewai>=0.28.0
crewai_tools>=0.1.6
supabase>=1.0.0
langchain-openai>=0.1.0
langchain-google-genai>=0.1.0
datamodel-code-generator>=0.25.2
python-jose[cryptography]>=3.3.0

# Letta Client (MemGPT)
letta-client

# Scheduler
APScheduler>=3.0.0,<4.0.0

# Database ORM
SQLAlchemy>=2.0.0,<2.1.0

# RSS Feed Parsing
feedparser>=6.0.0,<7.0.0

# Web3 for DEX interaction
web3>=6.0,<7.0